♻️ Refactor: create useLogin custom hook for Login component
